requirements.txt
Django==2.0.2
django-crispy-forms==1.7.2
django-crudbuilder==0.2.7
django-tables2==1.19.0
django-widget-tweaks==1.4.1
pytz==2017.3
six==1.11.0
